pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-confhub-cli"
version = "0.1.0"
description = "AI 에이전트 설정 중앙 관리 CLI 프레임워크"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "jinja2>=3.1.0",
]

[project.scripts]
nxs = "confhub.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pre-commit>=3.0",
    "ruff>=0.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["confhub"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["confhub"]
omit = ["tests/*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["confhub"]
